IPv6: Don't delete local address if addr_gen_mode is none
This can help us speed up VRF assignment, as when the interface is taken
down, the address won't be deleted and we won't run
rt_genid_bump_ipv6()/__fib6_flush_trees().

It is a costly operation as this goes over all of the nodes in all of
the trees in the netns.

Signed-off-by: Linik <ylinik@drivenets.com>