Do not invoke addrconf_verify_rtnl unnecessarily
Do not invoke costly `addrconf_verify_rtnl` if the added address
wouldn't need it, or affect the delayed_work timer.

For new/modified addresses, call "verify" only if added address has an
expiration, or if any temporary address was created.

This is done to account for a case where the new expiration time might
be sooner than the current delayed_work's expiration.

For deleted addresses, avoid calling verify at all:

If the address being deleted is not perishable, and thus does not affect
the delayed_work expiration, there is not point in going over the entire
table.

If the address IS perishable, but is not the soonest-to-be-expired
address, calling "verify" would not change the expiration, and would be
a very expensive nop.

If the address IS perishable, and IS the soonest-to-be-expired address,
calling or not-calling "verify" for a single address deletion is
equivalent in cost.

But calling "verify" immediately will result in a performance hit when
deleting many addresses.

Signed-off-by: Gilad Naaman <gnaaman@drivenets.com>